refactor: consolidate mock declarations in setup.ts
